+++ src/base/selection.cpp
@@ -132,15 +132,15 @@ bool Selection::leftClick( Position& pos, bool shift, bool ctrl )
 		m_firstClick          = pos;
 		m_firstClicked        = true;
 		QVariantMap actionMap = DB::selectRow( "Actions", m_action );
-		m_isFloor             = DB::select( "IsFloor", "Actions", m_action ).toBool();
-		m_isMulti             = DB::select( "Multi", "Actions", m_action ).toBool();
+		m_isFloor             = actionMap.value( "IsFloor" ).toBool();
+		m_isMulti             = actionMap.value( "Multi" ).toBool();
 		if ( m_action == "BuildItem" && DB::select( "IsContainer", "Items", m_item ).toBool() )
 		{
 			m_isMulti = true;
 		}
 
-		m_isMultiZ  = DB::select( "MultiZ", "Actions", m_action ).toBool();
-		m_canRotate = DB::select( "Rotate", "Actions", m_action ).toBool();
+		m_isMultiZ  = actionMap.value( "MultiZ" ).toBool();
+		m_canRotate = actionMap.value( "Rotate" ).toBool();
 		m_selection.push_back( QPair<Position, bool>( pos, testTileForJobSelection( pos ) ) );
 		if ( !m_isMulti && !shift )
 		{
@@ -172,10 +172,10 @@ void Selection::setAction( QString action )
 	m_action              = action;
 	QVariantMap actionMap = DB::selectRow( "Actions", m_action );
 
-	m_isFloor   = DB::select( "IsFloor", "Actions", m_action ).toBool();
-	m_isMulti   = DB::select( "Multi", "Actions", m_action ).toBool();
-	m_isMultiZ  = DB::select( "MultiZ", "Actions", m_action ).toBool();
-	m_canRotate = DB::select( "Rotate", "Actions", m_action ).toBool();
+	m_isFloor   = actionMap.value( "IsFloor" ).toBool();
+	m_isMulti   = actionMap.value( "Multi" ).toBool();
+	m_isMultiZ  = actionMap.value( "MultiZ" ).toBool();
+	m_canRotate = actionMap.value( "Rotate" ).toBool();
 }
 
 void Selection::updateSelection( Position& pos, bool shift, bool ctrl )
@@ -358,8 +358,8 @@ bool Selection::testTileForJobSelection( const Position& pos )
 		}
 		for ( auto req : required )
 		{
-			if ( Global::debugMode )
-				qDebug() << "test requirement: " << req << "...";
+			//if ( Global::debugMode )
+			//	qDebug() << "test requirement: " << req << "...";
 			switch ( m_reqMap.value( req ) )
 			{
 				case SEL_NONE:
@@ -479,8 +479,8 @@ bool Selection::testTileForJobSelection( const Position& pos )
 				break;
 			}
 
-			if ( Global::debugMode )
-				qDebug() << "passed.";
+			//if ( Global::debugMode )
+			//	qDebug() << "passed.";
 		}
 
 		QStringList forbidden = tm.value( "Forbidden" ).toString().split( "|" );
@@ -490,8 +490,8 @@ bool Selection::testTileForJobSelection( const Position& pos )
 		}
 		for ( auto forb : forbidden )
 		{
-			if ( Global::debugMode )
-				qDebug() << "test forbidden: " << forb << "...";
+			//if ( Global::debugMode )
+			//	qDebug() << "test forbidden: " << forb << "...";
 			if ( m_action.startsWith( "Build" ) && tile->flags & TileFlag::TF_OCCUPIED )
 			{
 				return false;
@@ -620,8 +620,8 @@ bool Selection::testTileForJobSelection( const Position& pos )
 					break;
 			}
 
-			if ( Global::debugMode )
-				qDebug() << "passed.";
+			//if ( Global::debugMode )
+			//	qDebug() << "passed.";
 		}
 	}